Code audit cleanup — stability, readability, and extensibility pass
Fixes:
- ROM::read/write bounds check now uses size_t comparison to avoid uint16_t
  truncation on ROMs larger than 64 KB
- size_t underflow guard in designer ROM preview for header-only PRG files
- BankedMemory::statusLine shows bytes (not "0 KB") for sub-1KB banks

Refactoring:
- drawMachineDesigner() (~630 lines) split into 8 focused helper methods
- Preset builders share new Machine::clearForPreset() for teardown boilerplate
- CIA6526 raw bit-mask literals replaced with named constants (ICR_SET_BIT,
  ICR_SOURCES, TOD_PM, TOD_HR_MASK, TOD_HR_WRITE_MASK)
- Designer message fields use new DesignerMsg struct with explicit isErr flag,
  replacing fragile rfind-based prefix matching for color selection
- Application.h Watchpoints section header added for consistency

+++ src/emulator/devices/ROM.cpp
@@ -31,12 +31,12 @@ bool ROM::loadFromFile(const std::string& path, size_t skipBytes) {
 }
 
 void ROM::write(uint16_t offset, uint8_t value) {
-    if (writable_ && offset < static_cast<uint16_t>(data_.size()))
+    if (writable_ && static_cast<size_t>(offset) < data_.size())
         data_[offset] = value;
 }
 
 uint8_t ROM::read(uint16_t offset) const {
-    if (offset < static_cast<uint16_t>(data_.size()))
+    if (static_cast<size_t>(offset) < data_.size())
         return data_[offset];
     return 0xFF;
 }
